requirements.txt
fastapi~=0.115.12
selenium~=4.30.0
python-dotenv~=1.0.1
uvicorn[standard]
redis>=5.0.1

unidecode~=1.3.8
requests~=2.32.3
pydantic~=2.10.6
beautifulsoup4~=4.13.3
starlette~=0.46.1